Exception handling - Client delete

+++ backend/src/main/java/com/devsuperior/desafio03/services/ClientService.java
@@ -55,6 +55,9 @@ public class ClientService {
 	
 	@Transactional
 	public void delete(Long id) {
+		if(!repository.existsById(id)) {
+			throw new ResourceNotFoundException("Recurso não encontrado");
+		}
 		repository.deleteById(id);
 	}
